Update CONDUCT.md (#31)
one typo corrected - in terms of the task of checking "consistency" - the Data Together CoC are nearly verbatim EDGI - the CivicTech TO is close - but is more vivid in terms of the wrath that will befall transgressors. Recurse is *very* different in terms of its formulation of the harassment and its effects. The Geek Feminism Wiki is focused on the "conference" - which may be too restrictive a space for Data Together's purposes, but their content is the *most* diverse & comprehensive and IMHO should be given more consideration.